requirements.txt
bokeh==2.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Flask==1.1.2
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
line-bot-sdk==1.16.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.2
packaging==20.3
Pillow==7.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
tornado==6.0.4
twstock==1.3.1
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
